src/app/app.component.spec.ts
import { TestBed, async } from '@angular/core/testing';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';
import { RouterTestingModule } from '@angular/router/testing';
import { AppComponent } from './app.component';

describe('AppComponent', () => {
  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [
        ReactiveFormsModule,
        FormsModule,
        RouterTestingModule
      ],
      declarations: [
        AppComponent
      ],
    }).compileComponents();
  }));

  it('should create the app', () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    expect(app).toBeTruthy();
  });

  it(`should have as title 'angular-unit-testing'`, () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    expect(app.title).toEqual('angular-unit-testing');
  });

  it('should render title', () => {
    const fixture = TestBed.createComponent(AppComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement;
    expect(compiled.querySelector('.content span').textContent).toContain('angular-unit-testing app is running!');
  });

  it('should render first name on checkbox selection', () => {
    const fixture = TestBed.createComponent(AppComponent);
    let component = fixture.componentInstance;
    let checkbox = component.contactForm.controls['showInput'];
    component.handleChange();
    checkbox.setValue(true);
    fixture.detectChanges();//check changes after value change.
    const compiled = fixture.nativeElement;
    expect(compiled.querySelector('.data').textContent).toContain('Add Data');
  });
});
